upd: rm sh script & move test
