+++ near-sdk-macros/Cargo.toml
@@ -26,10 +26,9 @@ serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 
 [dev-dependencies]
-near-abi = { version = "0.4.0", features = ["__chunked-entries"] }
 insta = { version = "1.31.0", features = ["yaml"] }
 prettyplease = { version = "0.2.15" }
-symbolic-debuginfo = "8.8"
+
 
 [features]
 abi = []
