+++ near-sdk-macros/src/lib.rs
@@ -643,63 +643,3 @@ pub fn derive_event_attributes(item: TokenStream) -> TokenStream {
         )
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use std::{collections::HashSet, fs, path::Path, process::Command};
-
-    // These methods are prepended to the contract internally, update this test list if they change
-    const PREPENDED_METHODS: [&str; 1] = ["contract_source_metadata"];
-    const PROJECT_PATH: &str = "../../near-sdk-rs/examples/adder";
-
-    #[test]
-    fn ensure_abi_for_prepended_functions() {
-        let contract_dir = Path::new(PROJECT_PATH);
-
-        let res = Command::new("bash")
-            .arg(contract_dir.join("abi.sh").to_str().expect("script path is valid"))
-            .output()
-            .unwrap();
-
-        assert!(
-            res.status.success(),
-            "failed to compile contract abi: {}",
-            String::from_utf8_lossy(&res.stderr)
-        );
-
-        let dylib_file = contract_dir.join(format!("res/libadder.{}", dylib_extension()));
-        assert!(dylib_file.exists(), "Dylib file should exist");
-
-        let dylib_file_contents =
-            fs::read(contract_dir.join(format!("res/libadder.{}", dylib_extension()))).unwrap();
-        let object = symbolic_debuginfo::Object::parse(&dylib_file_contents).unwrap();
-        let near_abi_symbols = object
-            .symbols()
-            .flat_map(|sym| sym.name)
-            .filter(|sym_name| sym_name.starts_with("__near_abi_"))
-            .collect::<HashSet<_>>();
-
-        // ensure methods are prepended
-        PREPENDED_METHODS.map(|method| {
-            assert!(
-                near_abi_symbols.iter().any(|f| f.eq(&format!("__near_abi_{}", method))),
-                "ABI should contain prepended method {}",
-                method
-            );
-        });
-    }
-
-    const fn dylib_extension() -> &'static str {
-        #[cfg(target_os = "linux")]
-        return "so";
-
-        #[cfg(target_os = "macos")]
-        return "dylib";
-
-        #[cfg(target_os = "windows")]
-        return "dll";
-
-        #[cfg(not(any(target_os = "macos", target_os = "linux", target_os = "windows")))]
-        compile_error!("Unsupported platform");
-    }
-}
